requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-dotenv
# Install the necessary library
google-cloud-aiplatform